+++ pkg/manager/test_mocks/simple.go
@@ -1,4 +1,3 @@
 package test_mocks
 
 // Simple test to see if file creation works
-