+++ CHANGES.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
-
+- Fixed custom DNS resolver configuration not being passed to Lego client
+- Fixed nil pointer dereference when verifying CNAME records in domain sharing scenarios
+- Restricted Lego to DNS-01 challenge type only (disabled HTTP-01 and TLS-ALPN-01)
 
 ## 0.5.2 - 2025-05-05